requirements.txt
psycopg2==2.9.9
django-stubs==4.2.7
scikit-learn==1.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.25
pyyaml==5.3
langchain==0.1.2
langchain-community==0.0.14
langchain-openai==0.0.3
langchain-experimental==0.0.49
chartjs==1.2